bool

+++ srcs/app/app_output.c
@@ -6,7 +6,7 @@
 /*   By: bconchit <bconchit@student.21-school.ru>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/27 21:07:32 by bconchit          #+#    #+#             */
-/*   Updated: 2020/07/21 14:12:14 by bconchit         ###   ########.fr       */
+/*   Updated: 2020/07/21 17:55:10 by bconchit         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,8 +27,8 @@ static void		output_steps(t_app *self)
 {
 	t_list_iter		*iter;
 	t_ant			*ant;
-	int				count;
-	int				step;
+	size_t			count;
+	size_t			step;
 
 	step = 0;
 	while (self->ants->count > 0)
